pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "librosa>=0.11.0",
    "numpy>=2.2.5",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "scipy>=1.15.2",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
